==> Hockey_Salary_Research/Script/3_Transform/step2_on_ice_data_clean.ipynb <==
# Taking a look at our On Ice data #

Here we will take a quick look at the player metrics collected

import pandas as pd
import numpy as np
from per_60_functions import per_60_to_totals, per_60_to_GP
import os

# Load bio data from directory "../../Data/NHL_PlayerData_NatrualStatTrick/oi" using pandas dataframe

directory = "../../Data/NHL_PlayerData_NaturalStatTrick/oi"

oi_df = pd.DataFrame()

for year in range(2007, 2025):
    filename = f'regSeason_allStrengths_allScores_oi_{year}.csv'
    df = pd.read_csv(f'{directory}/{filename}')
    # Concatenate dataframes with bio_df
    oi_df = pd.concat([oi_df, df])

oi_df.head(-5)

## The importance of our On Ice Data ##

This data is likely to be composed of very valuable features we will likely want to use for training our model, so we should do a fairly thorough job of cleaning and ensuring it is ready for proper analysis.

This data is composed of what are called "advanced metrics".

oi_df.info()

Here we see that given 15227 rows, not a single column has a null entry, which is fantastic.

# Use .describe() to get summary statistics of the data in a way that we view all columns

oi_df.describe()

# Converting Rates to Totals

The features of our data are all in rate form, and it occurs to me that it may be useful to utilize the totals for our features, instead of the rates. Here I will convert all features that are in a rate format "feature/60" back to simply "feature".

Something convenient that I noticed is that TOI is already in the correct format. As opposed to minutes:seconds it has already been converted to total minutes

per_60_columns = oi_df.filter(like='/60').columns

totals_df = per_60_to_totals(oi_df, per_60_columns)

# Merge oi_df with new_df on axis = 1
oi_df = pd.concat([oi_df, totals_df], axis=1)


gp_df = per_60_to_GP(oi_df, per_60_columns)

oi_df = pd.concat([oi_df, gp_df], axis=1)

# Extract all columns that end in '/60'

per_60_columns = oi_df.filter(like='/60')

per_GP_columns = oi_df.filter(like='/GP')

percentage_columns = oi_df.filter(like='%')

# Extract all columns that are not per 60 or per GP

non_per_columns = oi_df.drop(per_60_columns.columns, axis=1).drop(per_GP_columns.columns, axis=1).drop(percentage_columns.columns, axis=1)

final_data = pd.concat([non_per_columns, per_GP_columns, per_60_columns, percentage_columns], axis=1)
final_data.columns.tolist()

# Save this data to disk

dir_path = '../../Data/Warehouse/PlayerData/'
filename = 'oi.csv'

if not os.path.exists(dir_path):
    os.makedirs(dir_path)

final_data.to_csv(f'{dir_path}/{filename}', index=False)
