==> Hockey_Salary_Research/Script/3_Transform/step3_standard_data_clean.ipynb <==
# Taking a look at our Standard Player data #

Here we will take a quick look at the standard player metrics collected. This data includes things much of the traditional, and likely most important data we have. It includes metrics like goals, assists, points, +/-, and more.

import pandas as pd
from per_60_functions import per_60_to_totals, per_60_to_GP
import os

# Load bio data from directory "../../Data/NHL_PlayerData_NatrualStatTrick/oi" using pandas dataframe

directory = "../../Data/NHL_PlayerData_NaturalStatTrick/std"

std_df = pd.DataFrame()

for year in range(2007, 2025):
    filename = f'regSeason_allStrengths_allScores_std_{year}.csv'
    df = pd.read_csv(f'{directory}/{filename}')
    # Concatenate dataframes with bio_df
    std_df = pd.concat([std_df, df])

std_df.head(-5)

std_df.info()

Here we see that given 15227 rows, not a single column has a null entry, which is again fantastic.n Let's take a look at each of the columns to try to recognize if the values make sense.

# Use .describe() to get summary statistics of the data in a way that we view all columns

std_df.describe()

std_df.columns

std_df['Position'].unique()

# Converting Rates to Totals

The features of our data are all in rate form, and it occurs to me that it may be useful to utilize the totals for our features, instead of the rates. Here I will convert all features that are in a rate format "feature/60" back to simply "feature".

Something convenient that I noticed is that TOI is already in the correct format. As opposed to minutes:seconds it has already been converted to total minutes

per_60_columns = std_df.filter(like='/60').columns

totals_df = per_60_to_totals(std_df, per_60_columns)

# Merge std_df with new_df on axis = 1
std_df = pd.concat([std_df, totals_df], axis=1)


gp_df = per_60_to_GP(std_df, per_60_columns)

std_df = pd.concat([std_df, gp_df], axis=1)

# Extract all columns that end in '/60'

per_60_columns = std_df.filter(like='/60')

per_GP_columns = std_df.filter(like='/GP')

percentage_columns = std_df.filter(like='%')

# Extract all columns that are not per 60 or per GP

non_per_columns = std_df.drop(per_60_columns.columns, axis=1).drop(per_GP_columns.columns, axis=1).drop(percentage_columns.columns, axis=1)

final_data = pd.concat([non_per_columns, per_GP_columns, per_60_columns, percentage_columns], axis=1)
final_data.columns.tolist()

# Save this data to disk

dir_path = '../../Data/Warehouse/PlayerData/'
filename = 'std.csv'

if not os.path.exists(dir_path):
    os.makedirs(dir_path)

final_data.to_csv(f'{dir_path}/{filename}', index=False)
