==> Hockey_Salary_Research/Script/3_Transform/step5_salary_data_clean.ipynb <==
# Assessing Our Salary Data

import pandas as pd
import os

# Import all Y data from 2007-08 to 2023-24
Y_file_path = '../../Data/Lake/1B_Output_Salary/dataframe'

# Y_file_path contains csv files for each year from 2007-08 to 2023-24
salary_data = pd.DataFrame()
for year in range(2008, 2025):
    this_year = pd.read_csv(Y_file_path + '/season=' + str(year) + '.csv')
    salary_data = pd.concat([salary_data, this_year])

salary_data.shape

salary_data.columns.tolist()

# View players that show up twice in one year in our salary data
indices = salary_data[salary_data.duplicated(subset=['PLAYER', 'season'], keep=False)].sort_values(by=['PLAYER', 'POS', 'season']).index

# drop all records based on indices
salary_data.drop(index=indices, inplace=True)

salary_data.describe()

# Save salary data to the warehouse
outdir = '../../Data/Warehouse/SalaryData/'
outfile = 'salary_data.csv' 
if not os.path.exists(outdir):
    os.mkdir(outdir)
salary_data.to_csv(outdir + '/' + outfile, index=False)
